Changes issue #2354
1. In https://cloud.ibm.com/docs/services/sql-query?topic=sql-query-overview#database-locations update the paragraph:
"The <table name> part specifies the table that will be created in your database. It has the format <schema name>.<table name>. If you omit the part, the table is created in the schema of the "username" in the credentials of your database service instance – for a Db2 lite plan, this is the only schema that you have access to. The table name is case-preserving, so use upper case to match database defaults."

with the following changes:

"The <table name> part specifies the table that will be used when accessing your database. In case of INTO clause the table with this name will be created. It has the format <schema name>.<table name>. If you omit the <schema name> part, the table is looked for in the schema of the "username" in the credentials of your database service instance – for a Db2 lite plan, this is the only schema that you have access to. The table name is case-preserving, so use upper case to match database defaults."

and update the equivalent paragraph:

"The <table name> part specifies the table that will be created in your database. It has the format <schema name>.<table name>. If you omit the <schema name> part, the table is created in the schema of database user that has been created for the IBMid of the SQL query user. The table name is case-preserving, so use upper case to match database defaults."

with the following changes:

"The <table name> part specifies the table that will be used when accessing your database. In case of INTO clause the table with this name will be created. It has the format <schema name>.<table name>. If you omit the <schema name> part, the table is looked for in the schema of database user that has been created for the IBMid of the SQL query user. The table name is case-preserving, so use upper case to match database defaults."